refactor bitwise_endcoding into own module
