+++ lib/guardian/permissions/permission_encoding.ex
@@ -1,0 +1,7 @@
+defmodule Guardian.Permissions.PermissionEncoding do
+  @callback encode_permissions!(Guardian.Permissions.Bitwise.input_permissions() | nil) ::
+              Guardian.Permissions.Bitwise.t()
+  @callback decode_permissions(term) :: String.t()
+
+  @callback validate_permissions!(map) :: String.t()
+end
